:truck: Split dsp.rs into multiple files for readability
